Several fixes and cleanups (#42)
* Restructure towards 64-bit as a first choice. DVI/MDVI control algorithms updated to divide alpha and beta by num_states. Side by side implementations of sync updates in MDVI to ensure vectorized code is right. Fixed greedy_policy calculation in MDVI.

* DVI prediction now divides alpha and beta by num_states.

* Updates DVI-based prediction algorithms to divide alpha and beta by step size and tests params accordingly.

+++ src/differential_value_iteration/environments/structure_test.py
@@ -33,7 +33,7 @@ class StructureTest(parameterized.TestCase):
 
   @parameterized.parameters(*_MDPS)
   def test_create_mdp(self, mdp_constructor):
-    mdp = mdp_constructor(dtype=np.float32)
+    mdp = mdp_constructor(dtype=np.float64)
     self.assertTrue(mdp is not None)
 
   @parameterized.parameters(
